stats_service/requirements.txt
uvicorn==0.22.0
requests==2.28.2
httpx==0.23.3
passlib==1.7.4
email-validator==2.2.0
pytest==8.3.5
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.21.6
clickhouse-connect==0.8.17